+++ backend/src/test/java/com/woowacourse/f12/application/member/MemberServiceTest.java
@@ -7,8 +7,8 @@ import static com.woowacourse.f12.dto.CareerLevelConstant.SENIOR_CONSTANT;
 import static com.woowacourse.f12.dto.JobTypeConstant.BACKEND_CONSTANT;
 import static com.woowacourse.f12.dto.JobTypeConstant.ETC_CONSTANT;
 import static com.woowacourse.f12.support.InventoryProductFixtures.SELECTED_INVENTORY_PRODUCT;
-import static com.woowacourse.f12.support.KeyboardFixtures.KEYBOARD_1;
 import static com.woowacourse.f12.support.MemberFixtures.CORINNE;
+import static com.woowacourse.f12.support.ProductFixture.KEYBOARD_1;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
