+++ backend/src/test/java/com/woowacourse/f12/documentation/member/MemberDocumentation.java
@@ -3,8 +3,8 @@ package com.woowacourse.f12.documentation.member;
 import static com.woowacourse.f12.dto.CareerLevelConstant.JUNIOR_CONSTANT;
 import static com.woowacourse.f12.dto.JobTypeConstant.BACKEND_CONSTANT;
 import static com.woowacourse.f12.support.InventoryProductFixtures.SELECTED_INVENTORY_PRODUCT;
-import static com.woowacourse.f12.support.KeyboardFixtures.KEYBOARD_1;
 import static com.woowacourse.f12.support.MemberFixtures.CORINNE;
+import static com.woowacourse.f12.support.ProductFixture.KEYBOARD_1;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.willDoNothing;
