+++ backend/src/test/java/com/woowacourse/f12/presentation/member/MemberControllerTest.java
@@ -5,8 +5,8 @@ import static com.woowacourse.f12.dto.CareerLevelConstant.NONE_CONSTANT;
 import static com.woowacourse.f12.dto.JobTypeConstant.BACKEND_CONSTANT;
 import static com.woowacourse.f12.dto.JobTypeConstant.ETC_CONSTANT;
 import static com.woowacourse.f12.support.InventoryProductFixtures.SELECTED_INVENTORY_PRODUCT;
-import static com.woowacourse.f12.support.KeyboardFixtures.KEYBOARD_1;
 import static com.woowacourse.f12.support.MemberFixtures.CORINNE;
+import static com.woowacourse.f12.support.ProductFixture.KEYBOARD_1;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
